ISSUE #1431 - Renamed favourite metadata to starred metadata and merged the apikey test and starred metadata to the same file
